requirements.txt
flask>=3.0
openai>=1.0
html2image>=2.0
python-docx>=1.0
